+++ app/src/main/java/jp/co/penguin/gourmetsearch/data/entity/Area.kt
@@ -1,0 +1,6 @@
+package jp.co.penguin.gourmetsearch.data.entity
+
+data class Area (
+        var code: String? = null,
+        var name: String? = null
+)
