+++ app/src/main/java/jp/co/penguin/gourmetsearch/data/entity/Budget.kt
@@ -1,0 +1,7 @@
+package jp.co.penguin.gourmetsearch.data.entity
+
+data class Budget (
+        var code: String? = null,
+        var name: String? = null,
+        var average: String? = null
+)
