+++ app/src/main/java/jp/co/penguin/gourmetsearch/data/entity/Genre.kt
@@ -1,0 +1,7 @@
+package jp.co.penguin.gourmetsearch.data.entity
+
+data class Genre (
+        var code: String? = null,
+        var name: String? = null,
+        var catch: String? = null
+)
