+++ app/src/main/java/jp/co/penguin/gourmetsearch/data/entity/Urls.kt
@@ -1,0 +1,6 @@
+package jp.co.penguin.gourmetsearch.data.entity
+
+data class Urls (
+        var pc: String? = null,
+        var sp: String? = null
+)
